requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
scikit-learn>=1.0.0
xgboost>=1.5.0
optuna>=2.10.0
plotly>=5.3.0
joblib>=1.1.0
python-binance>=1.0.0
backtrader>=1.9.76.123
matplotlib>=3.5.0
imbalanced-learn>=0.10.0
ta>=0.10.0
PyYAML>=6.0
tqdm>=4.65.0
colorlog>=6.7.0
statsmodels>=0.13.0
scipy>=1.9.0
seaborn>=0.12.0
python-dateutil>=2.8.2
pyarrow>=14.0.1
fastparquet>=2023.10.1
optuna>=3.4.0
lightgbm>=4.1.0
catboost>=1.2.2
shap>=0.43.0
numba>=0.58.1
dask>=2023.12.1
ray>=2.8.0 